requirements.txt
cerberus==1.2
Flask==1.1.1
flask-restful==0.3.5
gunicorn==19.9.0
pymongo==3.8.0
pymodm==0.4.1
pyyaml==5.1
requests==2.7.0